Handle case when the app need to authenticate before make a call
